requirements.txt
agno
python-dotenv
uvicorn
fastapi
